@@ -1,43 +1,30 @@
 ﻿using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.TestHost;
-using Products.Api.Controllers;
 using Products.Api.IntegrationTests.Infrastructure;
+using Products.Api.Scenarios.S03MultipleHandlers;
 using Xunit;
 
 namespace Products.Api.IntegrationTests.Specs
 {
     [Collection(Collections.Api)]
-    public class S02CodeBasedControllerTests
+    public class S03MultipleHandlers
     {
         private readonly TestHostFixture _fixture;
 
-        public S02CodeBasedControllerTests(TestHostFixture fixture)
+        public S03MultipleHandlers(TestHostFixture fixture)
         {
             _fixture = fixture;
         }
 
         [Fact]
-        public async Task Only_Users_Over_The_Allowed_Age()
-        {
-            var builder = _fixture.Server.CreateHttpApiRequest<S02CodeBasedController>(
-                controller => controller.OnlyForOldPeople());
-
-            var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
-                .GetAsync();
-
-            await response.IsSuccessStatusCodeOrThrow();
-        }
-
-        [Fact]
         public async Task Users_With_BadgeNumber_And_Correct_Department_Can_Enter_The_Office()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S02CodeBasedController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.SeveralOptionsToEnterTheOffice());
 
             var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
+                .WithIdentity(Identities.Hugo)
                 .GetAsync();
 
             await response.IsSuccessStatusCodeOrThrow();
@@ -46,11 +33,11 @@ namespace Products.Api.IntegrationTests.Specs
         [Fact]
         public async Task Users_With_TempBadge_Can_Enter_The_Office()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S02CodeBasedController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.SeveralOptionsToEnterTheOffice());
 
             var response = await builder
-                .WithIdentity(Identities.RubenGarcia)
+                .WithIdentity(Identities.Ruben)
                 .GetAsync();
 
             await response.IsSuccessStatusCodeOrThrow();
