@@ -1,20 +1,19 @@
 ﻿using System.Net;
-using System.Net.Http;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.AspNetCore.TestHost;
-using Products.Api.Controllers;
 using Products.Api.IntegrationTests.Infrastructure;
+using Products.Api.Scenarios.S04MultipleHandlersBehavior;
 using Xunit;
 
 namespace Products.Api.IntegrationTests.Specs
 {
     [Collection(Collections.Api)]
-    public class S01BisIntroControllerTests
+    public class S04MultipleHandlersBehaviorTests
     {
         private readonly TestHostFixture _fixture;
 
-        public S01BisIntroControllerTests(TestHostFixture fixture)
+        public S04MultipleHandlersBehaviorTests(TestHostFixture fixture)
         {
             _fixture = fixture;
         }
@@ -22,11 +21,11 @@ namespace Products.Api.IntegrationTests.Specs
         [Fact]
         public async Task Values_Is_Authorized_When_Always_Succeed()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S01BisIntroController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.HandlersBehavior());
 
             var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
+                .WithIdentity(Identities.Hugo)
                 .GetAsync();
 
             response.StatusCode.Should().Be(HttpStatusCode.Forbidden);
