@@ -3,18 +3,18 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.AspNetCore.TestHost;
-using Products.Api.Controllers;
 using Products.Api.IntegrationTests.Infrastructure;
+using Products.Api.Scenarios.S05ResourceBasedPolicy;
 using Xunit;
 
 namespace Products.Api.IntegrationTests.Specs
 {
     [Collection(Collections.Api)]
-    public class S03ResourceBasedControllerTests
+    public class S05ResourceBasedPolicyTests
     {
         private readonly TestHostFixture _fixture;
 
-        public S03ResourceBasedControllerTests(TestHostFixture fixture)
+        public S05ResourceBasedPolicyTests(TestHostFixture fixture)
         {
             _fixture = fixture;
         }
@@ -22,11 +22,11 @@ namespace Products.Api.IntegrationTests.Specs
         [Fact]
         public async Task User_Gets_NotFound_For_Non_Existing_Product()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S03ResourceBasedController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.GetProduct("9999"));
 
             var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
+                .WithIdentity(Identities.Hugo)
                 .GetAsync();
 
             response.StatusCode.Should().Be(HttpStatusCode.NotFound);
@@ -35,11 +35,11 @@ namespace Products.Api.IntegrationTests.Specs
         [Fact]
         public async Task User_Can_Get_Owned_Products()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S03ResourceBasedController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.GetProduct("1234"));
 
             var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
+                .WithIdentity(Identities.Hugo)
                 .GetAsync();
 
             await response.IsSuccessStatusCodeOrThrow();
@@ -48,11 +48,11 @@ namespace Products.Api.IntegrationTests.Specs
         [Fact]
         public async Task User_Can_Not_Get_Not_Owned_Products()
         {
-            var builder = _fixture.Server.CreateHttpApiRequest<S03ResourceBasedController>(
+            var builder = _fixture.Server.CreateHttpApiRequest<Controller>(
                 controller => controller.GetProduct("4321"));
 
             var response = await builder
-                .WithIdentity(Identities.HugoBiarge)
+                .WithIdentity(Identities.Hugo)
                 .GetAsync();
 
             response.StatusCode.Should().Be(HttpStatusCode.Forbidden);
